Extract DoorLock.Dummy to GpioDummy
We’ll re-use this for the door bell
